+++ packages/sx05re/emuelec/config/emuelec/scripts/emuelecRunEmu.sh
@@ -28,10 +28,15 @@ TBASH="/usr/bin/bash"
 JSLISTENCONF="/emuelec/configs/jslisten.cfg"
 RATMPCONF="/tmp/retroarch/ee_retroarch.cfg"
 RATMPCONF="/storage/.config/retroarch/retroarch.cfg"
+
 set_kill_keys() {
+	
+# If jslisten is running we kill it first so that it can reload the config file. 
+[ pgrep -f "/emuelec/bin/jslisten" >/dev/null 2>&1 ] &&  killall jslisten
+
 	KILLTHIS=${1}
-    sed -i '/program=.*/d' ${JSLISTENCONF}
-	echo "program=\"/usr/bin/killall ${1}\"" >> ${JSLISTENCONF}
+	sed -i "2s|program=.*|program=\"/usr/bin/killall ${1}\"|" ${JSLISTENCONF}
+	
 	}
 
 # Make sure the /emuelec/logs directory exists
@@ -224,19 +229,19 @@ eval echo ${RUNTHIS} >> $EMUELECLOG
 if [[ "$KILLTHIS" != "none" ]]; then
 
 # We need to make sure there are at least 2 buttons setup (hotkey plus another) if not then do not load jslisten
-	KKBUTTON1=$(sed -n "s|^button1=\(.*\)|\1|p" "${JSLISTENCONF}")
-	KKBUTTON2=$(sed -n "s|^button2=\(.*\)|\1|p" "${JSLISTENCONF}")
+	KKBUTTON1=$(sed -n "3s|^button1=\(.*\)|\1|p" "${JSLISTENCONF}")
+	KKBUTTON2=$(sed -n "4s|^button2=\(.*\)|\1|p" "${JSLISTENCONF}")
 	if [ ! -z $KKBUTTON1 ] && [ ! -z $KKBUTTON2 ]; then
 		if [ ${KILLDEV} == "auto" ]; then
-			/emuelec/bin/jslisten &>> ${EMUELECLOG} &
+			/emuelec/bin/jslisten --mode hold &>> ${EMUELECLOG} &
 		else
-			/emuelec/bin/jslisten --device /dev/input/${KILLDEV} &>> ${EMUELECLOG} &
+			/emuelec/bin/jslisten --mode hold --device /dev/input/${KILLDEV} &>> ${EMUELECLOG} &
 		fi
 	fi
 fi
 
 # Only run fbfix on N2
-[[ ! -f "/ee_s905" ]] && /storage/.config/emuelec/bin/fbfix
+[[ "$EE_DEVICE" == "Amlogic-ng" ]] && /storage/.config/emuelec/bin/fbfix
 
 # Exceute the command and try to output the results to the log file if it was not dissabled.
 if [[ $LOGEMU == "Yes" ]]; then
@@ -248,13 +253,13 @@ else
 fi 
 
 # Only run fbfix on N2
-[[ ! -f "/ee_s905" ]] && /storage/.config/emuelec/bin/fbfix
+[[ "$EE_DEVICE" == "Amlogic-ng" ]] && /storage/.config/emuelec/bin/fbfix
 
 # Show exit splash
 ${TBASH} /emuelec/scripts/show_splash.sh exit
 
-# Kill jslisten, we don't need to but just to make sure 
-killall jslisten
+# Kill jslisten, we don't need to but just to make sure, dot not kill if using OdroidGoAdvance
+[[ "$EE_DEVICE" != "OdroidGoAdvance" ]] && killall jslisten
 
 # Just for good measure lets make a symlink to Retroarch logs if it exists
 if [[ -f "/storage/.config/retroarch/retroarch.log" ]]; then
